requirements.txt
asgiref==3.5.0
beautifulsoup4==4.9.3
Brlapi==0.8.2
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
cupshelpers==1.0
dbus-python==1.2.16
distlib==0.3.4
distro==1.5.0
distro-info==1.0
Django==3.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
filelock==3.4.2
gpg===1.14.0-unknown
html5lib==1.1
httplib2==0.18.1
idna==2.10
louis==3.16.0
lxml==4.6.3
Mako==1.1.3
MarkupSafe==1.1.1
menulibre==2.2.1
platformdirs==2.4.1
psutil==5.8.0
pycairo==1.16.2
pycups==2.0.1
pycurl==7.43.0.6
PyGObject==3.38.0
PyJWT==1.7.1
PySimpleSOAP==1.16.2
pysmbc==1.0.23
python-apt==2.2.1
python-debian==0.1.39
python-debianbts==3.1.0
pytz==2021.3
pyxdg==0.27
reportbug==7.10.3+deb11u1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.13.0
webencodings==0.5.1
xdg==5
